Text-based Dialogue Display
Began implementing a more in-depth and reusable text display system that pulls text from .json files to be displayed for the user. The text is displayed as if it is being typed but the whole system still needs a lot of work. For now, some issues are that each line of text is displayed on top of the last line, and that clicking inside of the game while text begins displaying can cause a crash. Aside from this, the display mode was updated from fullscreen to borderless window to allow for a more functional and efficient game and for a better user experience. Fullscreen was causing all other applications to shrink to annoyingly small sizes when the game was run, and alt-tabbing was slow. Borderless window fixed both issues.