Tweaks and Menu Layout
Hello! This time I went back to look at the text rendering and displaying system and worked out a way to display the text at a font size that is determined by the diagonal screen size of the display. This is done relative to a baseline of a 16pt font for a 1080p display, which is open to change in the future. For now, I removed the test that was being used to display text in the main.py file, and I moved the "typer" function into util.py (which used to be settings.py). main.py now uses the main loop to run the Menu class from menu.py to begin the game. It will also handle moving into the game from this menu when the player starts. It will have access to options, player saves, and exiting the game. For now, there is a basic layout of the menu with very basic functionality. Goals for next time are:

Develop sprites for the menu UI.
Develop functionality for rendering button sizes and menu sprites based on display size.
Make sure text is properly rendered before menu is displayed.
Make sure text appears on top of the buttons.
Make sure text is still being rendered for the menu based on display size.